refactor(step2): 공통 VO common 패키지로 이동
